+++ bob-esbuild.config.ts
@@ -1,0 +1,6 @@
+export const config: import('bob-esbuild').BobConfig = {
+  tsc: {
+    dirs: ['packages/*'],
+  },
+  verbose: true,
+};
