+++ packages/core/package.json
@@ -3,38 +3,38 @@
   "version": "0.0.4",
   "private": false,
   "scripts": {
+    "prepack": "pnpm prebuild && bob-esbuild build",
     "prebuild": "pnpm typecheck && shx rm -rf ./dist",
-    "build": "vite build",
+    "typecheck": "tsc --noEmit --target esnext",
     "dev:publish": "yalc publish --push",
     "lint": "pnpm lint:eslint && pnpm lint:prettier",
     "lint:eslint": "eslint \"src/**/*.{ts,tsx}\"",
     "lint:fix": "eslint \"src/**/*.{ts,tsx}\" --fix",
     "lint:prettier": "prettier --check \"src/**/*.{ts,tsx}\" *.js",
-    "lint:prettier:fix": "prettier --write \"src/**/*.{ts,tsx}\"",
-    "typecheck": "tsc --noEmit --target esnext"
+    "lint:prettier:fix": "prettier --write \"src/**/*.{ts,tsx}\""
   },
   "files": [
     "dist"
   ],
-  "main": "dist/index.css.js",
-  "module": "dist/index.css.js",
-  "import": "dist/index.css.js",
-  "types": "dist/types/index.css.d.ts",
+  "type": "module",
+  "main": "dist/index.js",
+  "module": "dist/index.mjs",
+  "types": "dist/index.d.ts",
   "exports": {
-    "./package.json": "./package.json",
-    "./index.css.js": {
-      "main": "dist/index.css.js",
-      "module": "dist/index.css.js",
-      "import": "dist/index.css.js",
-      "types": "dist/types/index.css.d.ts"
+    ".": {
+      "types": "./dist/index.d.ts",
+      "require": "./dist/index.js",
+      "main": "./dist/index.mjs",
+      "module": "./dist/index.mjs",
+      "import": "./dist/index.mjs"
     },
-    "./index.css": {
-      "main": "dist/index.css.js",
-      "module": "dist/index.css.js",
-      "import": "dist/index.css.js",
-      "types": "dist/types/index.css.d.ts"
-    },
-    "./main.css": "./dist/main.css"
+    "./*": {
+      "types": "./dist/*.d.ts",
+      "require": "./dist/*.js",
+      "main": "./dist/*.js",
+      "module": "./dist/*.mjs",
+      "import": "./dist/*.mjs"
+    }
   },
   "sideEffects": [
     "*.css.ts"
