+++ packages/core/src/styles.css.ts
@@ -1,4 +1,12 @@
-// @ts-nocheck
+import {
+  createGlobalTheme,
+  createGlobalThemeContract,
+  createTheme,
+  globalStyle,
+  style,
+  styleVariants,
+} from '@vanilla-extract/css';
+import { makeColors, makeContractRecord } from './utils';
 import {
   borderRadius,
   boxShadow,
@@ -9,38 +17,7 @@ import {
   lineHeights,
   sizes,
 } from '@nelson-ui/theme';
-import {
-  createGlobalTheme,
-  createGlobalThemeContract,
-  createTheme,
-  globalStyle,
-  style,
-  styleVariants,
-} from '@vanilla-extract/css';
 import { createSprinkles, defineProperties } from '@vanilla-extract/sprinkles';
-
-const getColor = (
-  color: keyof typeof colors.light | keyof typeof colors.dark,
-  theme: typeof colors.light | typeof colors.dark
-  // @ts-ignore
-) => colors.foundation[theme[color]];
-
-const makeColors = (_colors: typeof colors.light | typeof colors.dark) => {
-  const keys = Object.keys(_colors) as (keyof typeof colors.light)[];
-  const result = {} as any;
-  keys.forEach(key => {
-    result[key] = getColor(key, _colors);
-  });
-  return result as Record<keyof typeof colors.light | keyof typeof colors.dark, string>;
-};
-const makeContractRecord = <T>(record: T) => {
-  const keys = Object.keys(record);
-  const result = {} as any;
-  keys.forEach(key => {
-    result[key] = key;
-  });
-  return result as Record<keyof typeof record, keyof typeof record>;
-};
 
 const colorsContract = createGlobalThemeContract(
   {
