--- packages/core/tsconfig.build.json
@@ -1,16 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-  "compilerOptions": {
-    "noEmit": false,
-    "outDir": "./dist",
-    "rootDir": "./src",
-    "tsBuildInfoFile": "./tsconfig.build.tsbuildinfo",
-    "composite": true
-  },
-  "references": [
-    {
-      "path": "../theme/tsconfig.build.json"
-    }
-  ],
-  "include": ["src/**/*"]
-}
