+++ packages/react/src/box.tsx
@@ -1,5 +1,5 @@
 import { createBox } from '@dessert-box/react';
-import { atoms } from '@nelson-ui/core/index.css.js';
+import { atoms } from '@nelson-ui/core/styles.css';
 import { OwnProps } from '@radix-ui/react-polymorphic';
 
 export const Box = createBox({ atoms });
