+++ packages/react/src/theme-provider.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { themeClass } from '@nelson-ui/core/index.css.js';
+import { themeClass } from '@nelson-ui/core/styles.css';
 
 export const ThemeProvider: React.FC<{ theme: 'light' | 'dark' }> = ({
   children,
