--- packages/react/tsconfig.build.json
@@ -1,25 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-  "compilerOptions": {
-    "noEmit": false,
-    "outDir": "./dist",
-    "rootDir": "./src",
-    "baseUrl": ".",
-    "paths": {
-      "@nelson-ui/core/index.css.js": [
-        "../core"
-      ]
-    }
-  },
-  "references": [
-    {
-      "path": "../theme/tsconfig.build.json"
-    },
-    {
-      "path": "../core/tsconfig.build.json"
-    }
-  ],
-  "include": [
-    "src/**/*"
-  ]
-}
