--- packages/react/tsconfig.types.json
@@ -1,16 +1,0 @@
-{
-  "extends": "./tsconfig.build.json",
-  "compilerOptions": {
-    "noEmit": false,
-    "outDir": "./dist",
-    "declaration": true,
-    "declarationDir": "./dist/types"
-  },
-  "include": [
-    "src/**/*"
-  ],
-  "exclude": [
-    "node_modules",
-    "scripts"
-  ]
-}
