+++ tsconfig.build.json
@@ -1,6 +1,5 @@
 {
   "compilerOptions": {
-    "noEmit": true,
     "target": "ESNext",
     "module": "ESNext",
     "moduleResolution": "node",
